more API work
