+++ test/programmer.js
@@ -4,16 +4,13 @@ var Code = require('code');
 var Lab = require('lab');
 var lab = exports.lab = Lab.script();
 
-var bs2 = require('../');
-var revisions = bs2.revisions;
+var Programmer = require('../').Programmer;
 var hardware = require('../mock/hardware');
 
 var hi = new Buffer([0xFF, 0x00, 0x00, 0x00, 0x00, 0x30, 0xA0, 0xC7, 0x92, 0x66, 0x48, 0x13, 0x84, 0x4C, 0x35, 0x07, 0xC0, 0x4B]);
 var blink = new Buffer([0xfe, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x18, 0x14, 0x20, 0x8c, 0x0e, 0xd8, 0xc8, 0x7c, 0xff, 0x0e, 0x60, 0x4a, 0xae, 0xe8, 0x9f, 0x49, 0xc1, 0x50, 0xc3, 0x6f, 0x8d, 0xd1, 0x03, 0x07, 0xc0, 0x60]);
 
-lab.experiment('bs2', function () {
-
-  // TODO: this doesn't need all these tests, just do it in programmer
+lab.experiment('Programmer', function () {
 
   var transport;
 
@@ -24,11 +21,8 @@ lab.experiment('bs2', function () {
 
   lab.test('times out', function (done) {
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2
-    };
-    bs2.identify(options, function(error){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2' });
+    bs2.identify(function(error){
 
       Code.expect(error).to.exist();
       Code.expect(error.message).to.equal('BS2 did not respond. Check power, connection, or maybe this is not a BS2');
@@ -40,11 +34,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0xC8, 0xAD, 0xCE, 0x10]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2
-    };
-    bs2.identify(options, function(error){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2' });
+    bs2.identify(function(error){
 
       Code.expect(error).to.exist();
       Code.expect(error.message).to.equal('Incorrect Response: 200. Board might not be a BS2');
@@ -56,11 +47,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0xBE, 0xAD, 0xCE, 0x10]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2
-    };
-    bs2.identify(options, function(error, result){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2' });
+    bs2.identify(function(error, result){
 
       Code.expect(error).to.not.exist();
       Code.expect(result).to.deep.equal({name: 'BS2', version: '1.0'});
@@ -72,11 +60,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0xBE, 0xAD, 0xCE, 0x00]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2
-    };
-    bs2.identify(options, function(error, result){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2' });
+    bs2.identify(function(error, result){
 
       Code.expect(error).to.not.exist();
       Code.expect(result).to.deep.equal({name: 'BS2', version: '0'});
@@ -88,11 +73,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x65]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2e
-    };
-    bs2.identify(options, function(error, result){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2e' });
+    bs2.identify(function(error, result){
 
       Code.expect(error).to.not.exist();
       Code.expect(result).to.deep.equal({name: 'BS2e', version: '1.0'});
@@ -104,11 +86,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x02]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2e
-    };
-    bs2.identify(options, function(error){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2e' });
+    bs2.identify(function(error){
 
       Code.expect(error).to.exist();
       Code.expect(error.message).to.equal('Unknown: 2');
@@ -121,11 +100,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x58]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2sx
-    };
-    bs2.identify(options, function(error, result){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2sx' });
+    bs2.identify(function(error, result){
 
       Code.expect(error).to.not.exist();
       Code.expect(result).to.deep.equal({name: 'BS2sx', version: '1.0'});
@@ -137,11 +113,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x59]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2sx
-    };
-    bs2.identify(options, function(error, result){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2sx' });
+    bs2.identify(function(error, result){
 
       Code.expect(error).to.not.exist();
       Code.expect(result).to.deep.equal({name: 'BS2sx', version: '1.1'});
@@ -153,11 +126,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x60]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2sx
-    };
-    bs2.identify(options, function(error, result){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2sx' });
+    bs2.identify(function(error, result){
 
       Code.expect(error).to.not.exist();
       Code.expect(result).to.deep.equal({name: 'BS2sx', version: '1.2'});
@@ -169,11 +139,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x02]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2sx
-    };
-    bs2.identify(options, function(error){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2sx' });
+    bs2.identify(function(error){
 
       Code.expect(error).to.exist();
       Code.expect(error.message).to.equal('Unknown: 2');
@@ -185,11 +152,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x70]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2p
-    };
-    bs2.identify(options, function(error, result){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2p' });
+    bs2.identify(function(error, result){
 
       Code.expect(error).to.not.exist();
       Code.expect(result).to.deep.equal({name: 'BS2p24', version: '1.0'});
@@ -201,11 +165,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x71]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2p
-    };
-    bs2.identify(options, function(error, result){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2p' });
+    bs2.identify(function(error, result){
 
       Code.expect(error).to.not.exist();
       Code.expect(result).to.deep.equal({name: 'BS2p24', version: '1.1'});
@@ -217,11 +178,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x72]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2p
-    };
-    bs2.identify(options, function(error, result){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2p' });
+    bs2.identify(function(error, result){
 
       Code.expect(error).to.not.exist();
       Code.expect(result).to.deep.equal({name: 'BS2p24', version: '1.2'});
@@ -233,11 +191,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x73]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2p
-    };
-    bs2.identify(options, function(error, result){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2p' });
+    bs2.identify(function(error, result){
 
       Code.expect(error).to.not.exist();
       Code.expect(result).to.deep.equal({name: 'BS2p24', version: '1.3'});
@@ -249,11 +204,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x50]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2p
-    };
-    bs2.identify(options, function(error, result){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2p' });
+    bs2.identify(function(error, result){
 
       Code.expect(error).to.not.exist();
       Code.expect(result).to.deep.equal({name: 'BS2p40', version: '1.0'});
@@ -265,11 +217,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x51]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2p
-    };
-    bs2.identify(options, function(error, result){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2p' });
+    bs2.identify(function(error, result){
 
       Code.expect(error).to.not.exist();
       Code.expect(result).to.deep.equal({name: 'BS2p40', version: '1.1'});
@@ -281,11 +230,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x52]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2p
-    };
-    bs2.identify(options, function(error, result){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2p' });
+    bs2.identify(function(error, result){
 
       Code.expect(error).to.not.exist();
       Code.expect(result).to.deep.equal({name: 'BS2p40', version: '1.2'});
@@ -297,11 +243,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x53]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2p
-    };
-    bs2.identify(options, function(error, result){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2p' });
+    bs2.identify(function(error, result){
 
       Code.expect(error).to.not.exist();
       Code.expect(result).to.deep.equal({name: 'BS2p40', version: '1.3'});
@@ -313,11 +256,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x02]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2p
-    };
-    bs2.identify(options, function(error){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2p' });
+    bs2.identify(function(error){
 
       Code.expect(error).to.exist();
       Code.expect(error.message).to.equal('Unknown: 2');
@@ -329,11 +269,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x69]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2pe
-    };
-    bs2.identify(options, function(error, result){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2pe' });
+    bs2.identify(function(error, result){
 
       Code.expect(error).to.not.exist();
       Code.expect(result).to.deep.equal({name: 'BS2pe24', version: '1.0'});
@@ -345,11 +282,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x70]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2pe
-    };
-    bs2.identify(options, function(error, result){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2pe' });
+    bs2.identify(function(error, result){
 
       Code.expect(error).to.not.exist();
       Code.expect(result).to.deep.equal({name: 'BS2pe24', version: '1.1'});
@@ -361,11 +295,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x71]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2pe
-    };
-    bs2.identify(options, function(error, result){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2pe' });
+    bs2.identify(function(error, result){
 
       Code.expect(error).to.not.exist();
       Code.expect(result).to.deep.equal({name: 'BS2pe24', version: '1.2'});
@@ -377,11 +308,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x49]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2pe
-    };
-    bs2.identify(options, function(error, result){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2pe' });
+    bs2.identify(function(error, result){
 
       Code.expect(error).to.not.exist();
       Code.expect(result).to.deep.equal({name: 'BS2pe40', version: '1.0'});
@@ -393,11 +321,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x50]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2pe
-    };
-    bs2.identify(options, function(error, result){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2pe' });
+    bs2.identify(function(error, result){
 
       Code.expect(error).to.not.exist();
       Code.expect(result).to.deep.equal({name: 'BS2pe40', version: '1.1'});
@@ -409,11 +334,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x51]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2pe
-    };
-    bs2.identify(options, function(error, result){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2pe' });
+    bs2.identify(function(error, result){
 
       Code.expect(error).to.not.exist();
       Code.expect(result).to.deep.equal({name: 'BS2pe40', version: '1.2'});
@@ -425,11 +347,8 @@ lab.experiment('bs2', function () {
 
     transport.setData(new Buffer([0x02]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2pe
-    };
-    bs2.identify(options, function(error){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2pe' });
+    bs2.identify(function(error){
 
       Code.expect(error).to.exist();
       Code.expect(error.message).to.equal('Unknown: 2');
@@ -439,16 +358,13 @@ lab.experiment('bs2', function () {
 
   lab.test('bootload throws with non multiple of 18 byte data', function (done) {
 
-    var throws = function () {
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2' });
 
-      var options = {
-      transport: transport,
-      revision: revisions.bs2
-    };
-      bs2.bootload(options, new Buffer([0x00, 0x01, 0x02, 0x03]), function(){});
-    };
+    function invalidData () {
+      bs2.bootload(new Buffer([0x00, 0x01, 0x02, 0x03]), function(){});
+    }
 
-    Code.expect(throws).to.throw(Error, 'Data must be in multiples of 18 bytes');
+    Code.expect(invalidData).to.throw(Error, 'Data must be in multiples of 18 bytes');
     done();
   });
 
@@ -457,11 +373,8 @@ lab.experiment('bs2', function () {
     //send bs2 response bytes, then the success byte
     transport.setData(new Buffer([0xBE, 0xAD, 0xCE, 0x10, 0x00]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2
-    };
-    bs2.bootload(options, hi, function(error){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2' });
+    bs2.bootload(hi, function(error){
 
       Code.expect(error).to.not.exist();
       done();
@@ -473,11 +386,8 @@ lab.experiment('bs2', function () {
     //send bs2 response bytes, then 2 success bytes, 1 for each packet
     transport.setData(new Buffer([0xBE, 0xAD, 0xCE, 0x10, 0x00, 0x00]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2
-    };
-    bs2.bootload(options, blink, function(error){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2' });
+    bs2.bootload(blink, function(error){
 
       Code.expect(error).to.not.exist();
       done();
@@ -489,11 +399,8 @@ lab.experiment('bs2', function () {
     //send bs2 response bytes, then the error byte which should stop bootload at first 18 byte packet
     transport.setData(new Buffer([0xBE, 0xAD, 0xCE, 0x10, 0x01]));
 
-    var options = {
-      transport: transport,
-      revision: revisions.bs2
-    };
-    bs2.bootload(options, blink, function(error){
+    var bs2 = new Programmer({ transport: transport, revision: 'bs2' });
+    bs2.bootload(blink, function(error){
 
       Code.expect(error).to.exist();
       Code.expect(error.message).to.equal('Board nacked packet 0 with code: 1');
